# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/signup": {
      "post": {
        "operationId": "UserController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/auth/signin": {
      "post": {
        "operationId": "UserController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/auth": {
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Incidents API",
    "description": "",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "type": "http"
      }
    },
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван",
            "minLength": 3,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "password123",
            "minLength": 8,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "description": "Электронная почта пользователя",
            "example": "user@example.com",
            "minLength": 3,
            "maxLength": 100
          },
          "surname": {
            "type": "string",
            "description": "Фамилия пользователя",
            "example": "Иванов",
            "minLength": 3,
            "maxLength": 100
          },
          "recaptchaToken": {
            "type": "string",
            "description": "Токен reCAPTCHA для проверки",
            "example": "03AGdBq24..."
          }
        },
        "required": [
          "name",
          "password",
          "email",
          "surname",
          "recaptchaToken"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Электронная почта пользователя для входа в систему",
            "example": "user@example.com",
            "minLength": 3,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя для входа в систему",
            "example": "password123",
            "minLength": 8,
            "maxLength": 100
          },
          "recaptchaToken": {
            "type": "string",
            "description": "Токен reCAPTCHA для подтверждения, что запрос сделан реальным пользователем",
            "example": "03AGdBq24..."
          }
        },
        "required": [
          "email",
          "password",
          "recaptchaToken"
        ]
      },
...
    }
  }
}